+++ src/types.d.ts
@@ -1,8 +1,4 @@
-import { Readable } from 'svelte/store';
-
-import type { CustomModifiers } from './parser/types';
-
-export type LoadingStore = Readable<boolean> & { toPromise: () => Promise<void[]>, get: () => boolean };
+import type { ParserOptions } from '@sveltekit-i18n/parser-default';
 
 export type Loader = () => Promise<Record<any, any>>;
 
@@ -15,30 +11,12 @@ export type LoaderModule = {
   loader: Loader;
 };
 
-export type DotNotationInput = Record<string, any> | null | undefined | any;
-
-export type DotNotationOutput = Record<string, any>;
-
-export type ToDotNotation = (input: DotNotationInput, parentKey?: string) => DotNotationOutput;
-
-export type FetchTranslations = (loaders: LoaderModule[]) => Promise<Record<string, DotNotationOutput>>;
-
-export type TranslationFunction = (key: string, vars?: Record<any, any>) => string;
-
-export type LocalTranslationFunction = (locale: string, key: string, vars?: Record<any, any>) => string;
-
-export type ExtendedStore<T, Get = () => T, Store = Readable<T>> = Store & { get: Get };
-
 export type ConfigTranslations = { [locale: string]: Record<string, any> };
-
-export type Translations = { [locale: string]: Record<string, string> };
 
 export type Config = {
   loaders?: LoaderModule[];
   translations?: ConfigTranslations;
   initLocale?: string;
   fallbackLocale?: string;
-  customModifiers?: CustomModifiers;
+  parserOptions?: ParserOptions;
 };
-
-export type GetConfig = (...params: any) => Config;
