+++ tests/specs/index.spec.ts
@@ -3,7 +3,7 @@ import i18n from '../../src/index';
 import { CONFIG, TRANSLATIONS } from '../data';
 import { filterTranslationKeys } from '../utils';
 
-const { initLocale = '', loaders, customModifiers } = CONFIG;
+const { initLocale = '', loaders } = CONFIG;
 
 describe('i18n instance', () => {
   it('exports all properties and methods', () => {
@@ -142,7 +142,7 @@ describe('i18n instance', () => {
     );
   });
   it('`addTranslations` prevents duplicit `loading`', async () => {
-    const { addTranslations, loadTranslations, loading } = new i18n({ loaders, customModifiers });
+    const { addTranslations, loadTranslations, loading } = new i18n({ loaders });
 
     addTranslations(TRANSLATIONS);
     loadTranslations(initLocale);
@@ -160,7 +160,7 @@ describe('i18n instance', () => {
   it('does not initialize without `initLocale`', async () => {
     const { initialized, loadConfig } = new i18n();
 
-    await loadConfig({ loaders, customModifiers });
+    await loadConfig({ loaders });
     const $initialized = get(initialized);
 
     expect($initialized).toBe(false);
@@ -243,14 +243,14 @@ describe('i18n instance', () => {
   it('`loadTranslations` method works without route', async () => {
     const { initialized, loadConfig, loadTranslations } = new i18n();
 
-    await loadConfig({ loaders, customModifiers });
+    await loadConfig({ loaders });
     expect(get(initialized)).toBe(false);
 
     await loadTranslations(initLocale);
     expect(get(initialized)).toBe(true);
   });
   it('`loadTranslations` method works for given routes only', async () => {
-    const { loadTranslations, translations } = new i18n({ loaders, customModifiers });
+    const { loadTranslations, translations } = new i18n({ loaders });
     const url = '/path#hash?a=b&c=d';
     const keys = (loaders || []).filter(({ routes }) => routes?.includes(url)).map(({ key }) => key);
 
@@ -263,205 +263,5 @@ describe('i18n instance', () => {
     expect(get(translations)[initLocale]).toEqual(
       expect.objectContaining(TRANSLATIONS[initLocale]),
     );
-  });
-});
-
-describe('parser', () => {
-  it('returns a key string if not defined', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.undefined')).toBe('common.undefined');
-  });
-  it('key returns proper value', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.no_placeholder')).toBe('NO_PLACEHOLDER');
-  });
-  it('placeholders work', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.placeholder', { value: 'TEST_VALUE' })).toBe('VALUES: TEST_VALUE, TEST_VALUE, TEST_VALUE, TEST_VALUE');
-  });
-  it('placeholders in payload work', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.placeholder', { value: 'TEST_{{another}}', another: 'VALUE' })).toBe('VALUES: TEST_VALUE, TEST_VALUE, TEST_VALUE, TEST_VALUE');
-  });
-  it('default value works for placeholders', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.placeholder_default')).toBe('VALUES: DEFAULT_VALUE, DEFAULT_VALUE, DEFAULT_VALUE, DEFAULT_VALUE');
-  });
-  it('dynamic default works for placeholders', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.placeholder_unknown', { default: 'DYNAMIC_DEFAULT_VALUE' })).toBe('DYNAMIC_DEFAULT_VALUE');
-  });
-  it('placeholders containing escaped values work', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.placeholder_escaped', { 'pl:ace;holder': 'TEST \\{\\{VALUE\\}\\}' })).toBe('TEST {{VALUE}}');
-  });
-  it('`eq` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_eq', { value: 'option9' })).toBe('VALUES: DEFAULT VALUE, DEFAULT VALUE, DEFAULT VALUE, DEFAULT VALUE');
-    expect($t('common.modifier_eq', { value: 'option2' })).toBe('VALUES: VALUE2, VALUE2, VALUE2, VALUE2');
-    expect($t('common.modifier_eq')).toBe('VALUES: DEFAULT VALUE, DEFAULT VALUE, DEFAULT VALUE, DEFAULT VALUE');
-  });
-  it('`ne` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_ne', { value: 10 })).toBe('DEFAULT VALUE');
-    expect($t('common.modifier_ne', { value: 5 })).toBe('VALUE2');
-    expect($t('common.modifier_ne', { value: 15 })).toBe('VALUE2');
-    expect($t('common.modifier_ne')).toBe('VALUE2');
-  });
-  it('`lt` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_lt', { value: 10 })).toBe('DEFAULT VALUE');
-    expect($t('common.modifier_lt', { value: 5 })).toBe('VALUE2');
-    expect($t('common.modifier_lt')).toBe('DEFAULT VALUE');
-  });
-  it('`lte` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_lte', { value: 10 })).toBe('VALUE2');
-    expect($t('common.modifier_lte', { value: 5 })).toBe('VALUE2');
-    expect($t('common.modifier_lte')).toBe('DEFAULT VALUE');
-  });
-  it('`gt` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_gt', { value: 10 })).toBe('VALUE1');
-    expect($t('common.modifier_gt', { value: 15 })).toBe('VALUE2');
-    expect($t('common.modifier_gt')).toBe('DEFAULT VALUE');
-  });
-  it('`gte` modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_gte', { value: 10 })).toBe('VALUE2');
-    expect($t('common.modifier_gte', { value: 15 })).toBe('VALUE2');
-    expect($t('common.modifier_gte')).toBe('DEFAULT VALUE');
-  });
-  it('`number` modifier works', async () => {
-    const { t, locales, locale, loadConfig, loadTranslations } = new i18n();
-
-    await loadConfig(CONFIG);
-    const value = 123456.789;
-    const altLocale = locales.get().find((l) => l !== initLocale) || '';
-
-    expect(t.get('common.modifier_number', { value })).toBe(new Intl.NumberFormat(initLocale, { maximumFractionDigits: 2 }).format(value));
-
-    locale.set(altLocale);
-    await loadTranslations(altLocale);
-
-    expect(t.get('common.modifier_number', { value })).toBe(new Intl.NumberFormat(altLocale, { maximumFractionDigits: 2 }).format(value));
-  });
-  it('`date` modifier works', async () => {
-    const { t, loadConfig, loadTranslations, locale, locales } = new i18n();
-
-    await loadConfig(CONFIG);
-    const value = Date.now();
-    const altLocale = locales.get().find((l) => l !== initLocale) || '';
-
-    expect(t.get('common.modifier_date', { value })).toBe(new Intl.DateTimeFormat(initLocale, { dateStyle: 'medium', timeStyle: 'short' }).format(value));
-
-    locale.set(altLocale);
-    await loadTranslations(altLocale);
-
-    expect(t.get('common.modifier_date', { value })).toBe(new Intl.DateTimeFormat(altLocale, { dateStyle: 'medium', timeStyle: 'short' }).format(value));
-  });
-  it('`ago` modifier works', async () => {
-    const { t, loadConfig, loadTranslations, locale, locales } = new i18n();
-
-    await loadConfig(CONFIG);
-    const value = Date.now() - 1000 * 60 * 30;
-    const altLocale = locales.get().find((l) => l !== initLocale) || '';
-
-    expect(t.get('common.modifier_ago', { value })).toBe(new Intl.RelativeTimeFormat(initLocale).format(-30, 'minute'));
-
-    locale.set(altLocale);
-    await loadTranslations(altLocale);
-
-    expect(t.get('common.modifier_ago', { value })).toBe(new Intl.RelativeTimeFormat(altLocale).format(-30, 'minute'));
-  });
-  it('custom modifier works', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_custom', { data: 'TEST_STRING' })).toBe('TEST_STRING');
-  });
-  it('modifiers containing escaped values work', async () => {
-    const { t, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $t = t.get;
-
-    expect($t('common.modifier_escaped', { 'va:lue': 'option:1' })).toBe('VA;{{LUE}}:1');
-    expect($t('common.modifier_escaped', { 'va:lue': 'option:2' })).toBe('VA;{{LUE}}:2');
-    expect($t('common.modifier_escaped')).toBe('DEFAULT {{VALUE}};');
-  });
-  it('with user-defined locale works', async () => {
-    const { t, l, loadConfig } = new i18n();
-
-    await loadConfig(CONFIG);
-    const $l = l.get;
-    const $t = t.get;
-
-    const tests: Array<[string, any]> = [
-      ['common.undefined', undefined],
-      ['common.no_placeholder', undefined],
-      ['common.placeholder', { value: 'TEST_VALUE' }],
-      ['common.modifier_gte', { value: 10 }],
-      ['common.modifier_custom', { data: 'TEST_STRING' }],
-      ['common.modifier_escaped', { 'va:lue': 'option:2' }],
-    ];
-
-    tests.forEach((options) => {
-      expect($l(initLocale, ...options)).toBe($t(...options));
-    });
   });
 });
