+++ tests/utils/index.ts
@@ -4,3 +4,20 @@ export const filterTranslationKeys = (localeTranslation: any, keys: string[]) =>
 }),
 {},
 );
+
+export const useDefault = <T = any>(value: any, def:any = {}): T => value || def;
+
+type DotNotationInput = Record<string, any> | null | undefined | any;
+
+type DotNotationOutput = Record<string, any>;
+
+type ToDotNotation = (input: DotNotationInput, parentKey?: string) => DotNotationOutput;
+
+export const toDotNotation: ToDotNotation = (input, parentKey) => Object.keys(useDefault(input)).reduce((acc, key) => {
+  const value = input[key];
+  const outputKey = parentKey ? `${parentKey}.${key}` : `${key}`;
+
+  if (value && typeof value === 'object') return ({ ...acc, ...toDotNotation(value, outputKey) });
+
+  return ({ ...acc, [outputKey]: value });
+}, {});
